Add case for multiple targets for the same ECU ID.
The Uptane standard indicates that this is invalid, and indeed, it is
unclear how an ECU should react in such a case.

I tried also adding a test with one target with two identical ECU IDs,
but the dictionary and json dump logic makes that near impossible. It's
basically meaningless anyway, since if it's redundant, it's redundant,
and if the HW ID is different, it wouldn't match anyway.

Signed-off-by: Patrick Vacek <patrickvacek@gmail.com>